requirements.txt
pygame
pyyaml
six
